% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding.r
\name{req}
\alias{req}
\title{Response at equilibrium}
\usage{
req(conc, rmax, kd)
}
\arguments{
\item{conc}{Analyte concentration.}

\item{rmax}{Maximum response.}

\item{kd}{Equilibrium dissociation constant.}
}
\description{
Returns the response value at equilibrium from concentration, Rmax and KD.
}
\examples{
req(6e-7,1.2,6e-7)
}
