% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding.r
\name{tteq}
\alias{tteq}
\title{Time to Equilibrium}
\usage{
tteq(conc, kon, koff, theta = 0.95)
}
\arguments{
\item{conc}{Analyte concentration.}

\item{kon}{Kon binding constant.}

\item{koff}{Koff binding constant.}

\item{theta}{Default 0.95.}
}
\description{
Returns the time taken to reach 95\% equilibrium.
}
\examples{
tteq(6e-7,20000,0.01)
}
